+++ src/lib/SX127xDriver/SX127x.cpp
@@ -204,8 +204,7 @@ void ICACHE_RAM_ATTR SX127xDriver::SetFrequencyHz(uint32_t freq)
   currFreq = freq;
   SetMode(SX127x_OPMODE_STANDBY);
 
-  //int32_t FRQ = ((uint32_t)((double)freq / (double)FREQ_STEP));
-  int32_t FRQ = ((uint32_t)((uint32_t)freq / (uint32_t)FREQ_STEP));
+  int32_t FRQ = ((uint32_t)((float)freq / (float)FREQ_STEP));
 
   uint8_t FRQ_MSB = (uint8_t)((FRQ >> 16) & 0xFF);
   uint8_t FRQ_MID = (uint8_t)((FRQ >> 8) & 0xFF);
